amg_time/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-filter==25.1
drf-access-policy==1.5.0
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.10.1
pyparsing==3.2.1
rest_condition==1.0.3
sqlparse==0.5.3
timedelta==2020.12.3
typing_extensions==4.12.2
